REFATCTOR(App): move routes to routes folder
